Remove HttpErrorFunctions, replaced by HttpReads impls
